# docs/swagger.yaml
basePath: /api/
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.ShowPlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      id:
        type: integer
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
    type: object
...
  models.UpdatePlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
      team:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Fantasy football manager microservice.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fantasy football manager API
  version: "1.0"
paths:
  /players/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a player
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete a player
      tags:
      - Players
    get:
      consumes:
      - application/json
      description: Get a player by ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowPlayer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Show a player
      tags:
      - Players
    patch:
      consumes:
      - application/json
      description: Update a player
      parameters:
      - description: Update player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePlayer'
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a player
      tags:
      - Players
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"